🛠️ Componentize each booking form step
